Isolate inspection data thread
